+++ example/server.js
@@ -147,11 +147,11 @@ function init (config, callback) {
             }
           }, {
             method: 'GET',
-            path: '/getChalAndMatchingOrgs',
+            path: '/getMatchingOrgs',
             handler: function (request, reply) {
               var chalId = request.query.chal_id;
 
-              request.server.methods.pg.challenges.getChalAndMatchingOrgs(chalId, function (error, response) { // eslint-disable-line
+              request.server.methods.pg.challenges.getMatchingOrgs(chalId, function (error, response) { // eslint-disable-line
                 Hoek.assert(!error, 'toggleActive failed');
                 reply(response);
               });
