+++ lib/formatters/index.js
@@ -2,10 +2,10 @@
 
 var getById = require('./getById.js');
 var getByTag = require('./getByTag.js');
-var getChalAndMatchingOrgs = require('./getChalAndMatchingOrgs.js');
+var getMatchingOrgs = require('./getMatchingOrgs.js');
 
 module.exports = {
   getById: getById,
   getByTag: getByTag,
-  getChalAndMatchingOrgs: getChalAndMatchingOrgs
+  getMatchingOrgs: getMatchingOrgs
 };
