+++ lib/index.js
@@ -82,13 +82,13 @@ function register (server, options, next) {
     });
   };
 
-  var getChalAndMatchingOrgs = function (chal_id, cb) {
-    query(queries.getChalAndMatchingOrgs(chal_id), pool, function (queryErr, res) {
+  var getMatchingOrgs = function (chal_id, cb) {
+    query(queries.getMatchingOrgs(chal_id), pool, function (queryErr, res) {
       if (queryErr) {
         return cb(queryErr);
       }
 
-      return cb(null, formatters.getChalAndMatchingOrgs(res.rows));
+      return cb(null, formatters.getMatchingOrgs(res.rows));
     });
   };
 
@@ -104,7 +104,7 @@ function register (server, options, next) {
     server.method('pg.challenges.getByTag', getByTag);
     server.method('pg.challenges.checkEditable', checkEditable);
     server.method('pg.challenges.toggleActive', toggleActive);
-    server.method('pg.challenges.getChalAndMatchingOrgs', getChalAndMatchingOrgs);
+    server.method('pg.challenges.getMatchingOrgs', getMatchingOrgs);
 
     return next();
   });
