+++ lib/queries/index.js
@@ -7,7 +7,7 @@ var getByTag = require('./getByTag.js');
 var getAllActive = require('./getAllActive.js');
 var checkEditable = require('./checkEditable.js');
 var toggleActive = require('./toggleActive.js');
-var getChalAndMatchingOrgs = require('./getChalAndMatchingOrgs.js');
+var getMatchingOrgs = require('./getMatchingOrgs.js');
 
 module.exports = {
   add: add,
@@ -17,5 +17,5 @@ module.exports = {
   getAllActive: getAllActive,
   checkEditable: checkEditable,
   toggleActive: toggleActive,
-  getChalAndMatchingOrgs: getChalAndMatchingOrgs
+  getMatchingOrgs: getMatchingOrgs
 };
