+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pg-challenges",
-  "version": "2.0.16",
+  "version": "2.0.17",
   "description": "Reusable challenges-system build with Postgres and Hapi.",
   "main": "lib/index.js",
   "scripts": {
